Move ObjectTree from components to common
