+++ assets/js/common/ObjectTree/ObjectTreeIcon.jsx
