+++ assets/js/components/ExecutionResults/CheckResultDetail/FactValue.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ObjectTree from '@components/ObjectTree';
+import ObjectTree from '@common/ObjectTree';
 
 function FactValue({ className, data }) {
   if (data === null) {
